Try move ssh fixture to classes.
