+++ tests/tests_integration/base.py
@@ -6,7 +6,6 @@ import subprocess
 import warnings
 
 import pytest
-import ssh_test_utils
 import test_utils
 
 TEST_PROJECT_NAME = "test_project"
@@ -75,17 +74,3 @@ class BaseTest:
         test_utils.delete_project_if_it_exists(project_name)
         yield project_name
         test_utils.delete_project_if_it_exists(project_name)
-
-    @pytest.fixture(
-        scope="class",
-    )
-    def setup_ssh_container(self):
-        # Annoying session scope does not seem to actually work
-        container_name = "running_ssh_tests"
-        ssh_test_utils.setup_ssh_container(container_name)
-        yield
-
-        sudo = "sudo " if platform.system() == "Linux" else ""
-
-        subprocess.run(f"{sudo}docker stop {container_name}", shell=True)
-        subprocess.run(f"{sudo}rm {container_name}", shell=True)
